Introduce FileSystemCache
This patch adds the structure that represents in-memory image of the
file system, together with the utilities for iterating over it.

On the creation, FileSystemCache will take a snapshot of a directory and
will create the initial image. After that, it's the user's responsiblity
to use `addNewFile` to keep the FileSystemCache fresh.